ai/requirements.txt
requests
python-docx
markdown-it-py